first fix and add Readme

+++ backend/users/views.py
@@ -1,45 +1,47 @@
 from rest_framework import status
-from rest_framework.generics import ListAPIView, get_object_or_404
+from rest_framework.decorators import action
+from rest_framework.generics import get_object_or_404
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
-from rest_framework.views import APIView
 
-from recipes.pagination import PageLimitPagination
+from recipes.mixins import CustomUserMixin
 from .models import Follow, User
 from .serializers import FollowListSerializer, FollowSerializer
 
 
-class FollowView(APIView):
-    """
-    APIView для добавления и удаления подписки на автора
-    """
-    serializer_class = FollowSerializer
-    permission_classes = [IsAuthenticated]
-    pagination_class = PageLimitPagination
+class FollowViewSet(CustomUserMixin):
+    queryset = User.objects.all()
+    serializer_class = FollowListSerializer
 
-    def post(self, request, id):
-        data = {'user': request.user.id, 'following': id}
-        serializer = FollowSerializer(data=data, context={'request': request})
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
+    @action(detail=True,
+            methods=["POST", "DELETE"],
+            permission_classes=[IsAuthenticated])
+    def subscribe(self, request, id):
+        if request.method == "POST":
+            data = {'user': request.user.id, 'following': id}
+            serializer = FollowSerializer(
+                data=data,
+                context={'request': request}
+            )
+            serializer.is_valid(raise_exception=True)
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
 
-    def delete(self, request, id):
         user = request.user
         following = get_object_or_404(User, id=id)
         follow = get_object_or_404(Follow, user=user, following=following)
         follow.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
-
-class FollowListView(ListAPIView):
-    """
-    APIView для просмотра подписок пользователя.
-    """
-    serializer_class = FollowListSerializer
-    permission_classes = [IsAuthenticated]
-    pagination_class = PageLimitPagination
-
-    def get_queryset(self):
+    @action(detail=False,
+            methods=["GET"],
+            permission_classes=[IsAuthenticated])
+    def subscriptions(self, request):
         user = self.request.user
-        return User.objects.filter(following__user=user)
+        queryset = User.objects.filter(following__user=user)
+        serializer = self.get_serializer(
+            queryset,
+            many=True,
+            context={'request': request}
+        )
+        return Response(serializer.data)
